feat(Drawer): introduce useToolbarData hook for managing toolbar state and simplify story implementations
